apps/qa_app/requirements.txt
langchain
python-dotenv
Openai
Streamlit